component folders created, styles movedЭ
