+++ src/components/BlockEditor/EditorCancel.jsx
@@ -1,0 +1,7 @@
+export const EditorCancel = () => {
+    return (
+        <button className="editor-cancel">
+            <img src="src/img/cancel-block-icon.svg" alt="cancel"/>
+        </button>
+    );
+};
