+++ src/components/BlockEditor/EditorSave.jsx
@@ -1,0 +1,7 @@
+export const EditorSave = () => {
+    return (
+        <button className="editor-save">
+            <img src="src/img/save-block-icon.svg" alt="save"/>
+        </button>
+    );
+};
