+++ src/components/Search/SearchCompany.jsx
@@ -1,0 +1,7 @@
+export const SearchCompany = ({ iconUrl }) => {
+    return (
+        <div className="search-company">
+            <img src={iconUrl} alt="company-icon" />
+        </div>
+    );
+};
