+++ src/components/Settings/Settings.jsx
@@ -1,0 +1,8 @@
+import './settings.css';
+import { SettingsIcon } from './SettingsIcon';
+
+export const Settings = () => {
+    return (
+        <SettingsIcon />
+    );
+};
